Rename RechartsAnimation to AnimationStateMachine
