+++ scripts/snapshots/es6Files.txt
@@ -3,9 +3,9 @@
   "es6/animation/AnimatedItems.js",
   "es6/animation/AnimationController.js",
   "es6/animation/AnimationControllerImpl.js",
+  "es6/animation/AnimationStateMachine.js",
   "es6/animation/CSSTransitionAnimate.js",
   "es6/animation/JavascriptAnimate.js",
-  "es6/animation/RechartsAnimation.js",
   "es6/animation/easing.js",
   "es6/animation/matchBy.js",
   "es6/animation/timeoutController.js",
