+++ scripts/snapshots/libFiles.txt
@@ -3,9 +3,9 @@
   "lib/animation/AnimatedItems.js",
   "lib/animation/AnimationController.js",
   "lib/animation/AnimationControllerImpl.js",
+  "lib/animation/AnimationStateMachine.js",
   "lib/animation/CSSTransitionAnimate.js",
   "lib/animation/JavascriptAnimate.js",
-  "lib/animation/RechartsAnimation.js",
   "lib/animation/easing.js",
   "lib/animation/matchBy.js",
   "lib/animation/timeoutController.js",
