+++ scripts/snapshots/typesFiles.txt
@@ -3,9 +3,9 @@
   "types/animation/AnimatedItems.d.ts",
   "types/animation/AnimationController.d.ts",
   "types/animation/AnimationControllerImpl.d.ts",
+  "types/animation/AnimationStateMachine.d.ts",
   "types/animation/CSSTransitionAnimate.d.ts",
   "types/animation/JavascriptAnimate.d.ts",
-  "types/animation/RechartsAnimation.d.ts",
   "types/animation/easing.d.ts",
   "types/animation/matchBy.d.ts",
   "types/animation/timeoutController.d.ts",
