+++ test/animation/mockAnimationController.ts
@@ -1,6 +1,6 @@
 import { AnimationController } from '../../src/animation/AnimationController';
 import { CancelableTimeout, TimeoutController } from '../../src/animation/timeoutController';
-import { RechartsAnimation } from '../../src/animation/RechartsAnimation';
+import { AnimationStateMachine } from '../../src/animation/AnimationStateMachine';
 
 /**
  * Animation controller that does not automatically complete when reaching 100%.
@@ -10,7 +10,7 @@ import { RechartsAnimation } from '../../src/animation/RechartsAnimation';
  */
 export const mockAnimationController: AnimationController = <T, E>(
   timeoutController: TimeoutController,
-  animationHandle: RechartsAnimation<T, E>,
+  animationHandle: AnimationStateMachine<T, E>,
   listener: (progress: T) => void,
 ): CancelableTimeout => {
   let cancellable: CancelableTimeout | undefined;
